% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSeqsWithMultinomialModel.R
\name{generateSeqsWithMultinomialModel}
\alias{generateSeqsWithMultinomialModel}
\title{Aleatorizacion de una secuencia.}
\usage{
generateSeqsWithMultinomialModel(inputsequence, X)
}
\arguments{
\item{inputsequence}{objeto tipo cadena de carácteres con la secuencia a aleatorizar.}

\item{X}{obtejo numerico. Numero de veces que se desea aleatorizar la secuencia.}
}
\value{
\code{generateSeqsWithMultinomialModel} devuelve un vector de longitud X con elementos de tipo
caracter, que son las secuencias ordenadas al azar.
}
\description{
La funcion \code{generateSeqsWithMultinomialModel} genera un numero determinado (por ejemplo, 1000) de 
secuencias aleatorias de nucleotidos (si la secuencia es ADN) o de aminoacidos (si la secuencia es una
proteina) de una longitud determinada mediante un modelo multinomial.
}
\examples{

generateSeqsWithMultinomialModel("PAWHEAE",100)
generateSeqsWithMultinomialModel("ATTGGCACT",10)

}
\references{
\url{https://a-little-book-of-r-for-bioinformatics.readthedocs.io/en/latest/src/chapter4.html}
}
