% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_alignment.R
\name{seq_alignment}
\alias{seq_alignment}
\title{Sequence alignment and statistical significance.}
\usage{
seq_alignment(
  seq1,
  seq2,
  seq_type = c("protein", "dna"),
  seq_align = c("local", "global"),
  mat,
  gap,
  N,
  shuff = c(1, 2)
)
}
\arguments{
\item{seq1}{Secuencia 1 en formato Fasta.}

\item{seq2}{Secuencia 2 en formato Fasta.}

\item{seq_type}{Tipo de secuencia: "protein" o "dna"}

\item{seq_align}{Tipo de alineamiento: "local" o "global".}

\item{mat}{Matriz de substitución: PAMn, BLOSUMn, … Ver \code{data(package="Biostrings")} para ver todas las opciones disponibles.}

\item{gap}{Puntuación Gap: Open penalty, extended penalty.}

\item{N}{Numero de replicas.}

\item{shuff}{Secuencia donde se hace shuffling: 1 o 2.}
}
\value{
\code{seq_alignment} devuelve un resultado gráfico (histograma de las puntiaciones estimadas) y un resultado numérico.
}
\description{
La funcion \code{seq_alignment} muestra la significacion estadistica de un alineamiento
por parejas.
}
\examples{

seq_alignment(seq1 = "P0DP27.fa", seq2 = "Q9N1R0.fa", seq_type ="protein",
              seq_align = "local", mat = data("BLOSUM62"), gap = c(-12, -2), N = 100, shuff = 1)
seq_alignment(seq1 = "gi32121095_N_0.fa", seq2 = "gi32121095_N_1.fa", seq_type ="dna",
              seq_align = "global", mat = data("BLOSUM6100"), gap = c(-8, -1), N = 2000, shuff)
seq_alignment(seq1 = "P69905.fa", seq2 = "P01942.fa", seq_type ="protein",
              seq_align = "global", mat = data("BLOSUM62"), gap = c(10, 0.5), N =1000, shuff=1)

}
\references{
\url{https://a-little-book-of-r-for-bioinformatics.readthedocs.io/en/latest/src/chapter4.html}
}
