\name{seqBio-package}
\alias{seqBio-package}
\alias{seqBio}
\docType{package}
\title{
\packageTitle{seqBio}
}
\description{
\packageDescription{seqBio}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{seqBio}
\packageIndices{seqBio}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{seqBio}

Maintainer: \packageMaintainer{seqBio}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
