% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{seqBio}
\alias{seqBio}
\title{seqBio: A sequence alginment package.}
\description{
El paquete seqBio contiene una funcion llamada \code{seq_alignment()}.
}
\section{Object}{

La funcion \code{seq_alignment()} es util para estudiar la significacion estadistica de un alineamiento por parejas.
Para mas informacion ir a \code{?seq_alignment} o bien \code{browseVignettes('seqBio')}.
}

\author{
Laura Julia Melis.

Curs de Fonaments de Bioinformatica.

Master en Estadistica i Investigacio Operativa, Universitat de Barcelona.
}
