% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task2.R
\name{generateSeqsWithMultinomialModel}
\alias{generateSeqsWithMultinomialModel}
\title{Muestra la significación estadística de un alineamiento por parejas.}
\usage{
generateSeqsWithMultinomialModel(inputsequence, X)
}
\arguments{
\item{seq1}{Una secuencia en formato fasta}

\item{seq2}{Otra secuencia en formato fasta}

\item{tipos}{Tipo de secuencia(Proteina o ADN)}

\item{tipoal}{Tipo de alineamiento(“local” o “global”)}

\item{M}{Matriz de substitución(PAMn, BLOSUMn,...)}

\item{puntugap}{Gap: Open penalty, extended penalty}

\item{replcas}{Número de replicas}

\item{shuf}{Secuencia donde se hace shuffling: 1 o 2}
}
\value{
el histograma con los scores
}
\description{
Dados dos secuencias, hace shuffling en una de ellas y dibujamos
un histograma con el numero de scores, además de devolver una descripción.
}
