% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{seqBio}
\alias{seqBio}
\title{seqBio: A sequence alginment package.}
\description{
El paquete seqBio contiene una función llamada \code{seq_alignment()}. Esta función es 
útil para estudiar la significación estadística de un alineamiento por parejas.
}
\section{Class object}{

\code{linreg()} calculates various statistics related to the linear regression model and contains multiple methods to obtain the regression coefficients or the predicted values, among others.
}

\author{
Laura Julia Melis.

Curs de Fonaments de Bioinformàtica.

Master en Estadística i Investigació Operativa, Universitat de Barcelona.
}
