% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{seqBio}
\alias{seqBio}
\title{seqBio: A sequence alginment package.}
\description{
El paquete seqBio contiene dos funciones: \code{seq_alignment()} y \code{generateSeqsWithMultinomialModel}.
}
\section{Objects}{

La funcion \code{seq_alignment()} es util para estudiar la significacion estadistica de un alineamiento por parejas.
Para mas informacion dobre esta funcion ir a \code{?seq_alignment}.

La funcion \code{generateSeqsWithMultinomialModel} genera un numero determinado de secuencias aleatorias 
(de nucleotidos o de aminoacidos) de una longitud determinada mediante un modelo multinomial. Para mas
 informacion dobre esta funcion ir a \code{?generateSeqsWithMultinomialModel}.

Si se desea conocer mas sobre estas funciones a traves de un informe dinamico, 
ejecutar el codigo: \code{browseVignettes('seqBio')}.
}

\author{
Laura Julia Melis.

Curs de Fonaments de Bioinformatica.

Master en Estadistica i Investigacio Operativa, Universitat de Barcelona.
}
